Adaptaions for zsh
* Patched loader for shared functions as `${BASH_SOURCE}` isn't a thing in
zsh

* Created hook for `oh-my-zsh` plugin system to call

This changes gcp to gcpro and gcd to gcdns because of collisions with
the common oh-my-zsh git alias

Also the loader.zsh is converted to a zsh and it now gets the correct
path. The bash version just needs BASH_SOURCE but this is not available
with zsh and is not correct as we want the script directory not the
current directory. and the plugin.zsh is modified to point to that zsh.

This also splits the completion reading into complete.sh so all
loader.sh and loader.zsh do is to use different methods to determine the
plugin path

README is also updated